Introducing optional use of JSON transfer objects.
Added transfer objects for most relevant data structures. To be further refined as we move forward.
These rely on GSon's capabilities of easily converting Java types to/from JSON. The defaults were slightly extended to allow for use of JsonObject for certain transfer object members, which enables partial mapping of structures (e.g., map only the top-level structure, but obtain some complex proeprty of this structure as JsonObject, for an example see handling of change commands in ModelSessionRestController).

git-svn-id: http://emeafrazerg/svn/ipp/product/trunk/stardust/ui.web@59086 8100b5e0-4d52-466c-ae9c-bdeccbdeaf6b