Optionally apply commands in two phases to allow for enhanced concurrency.
The optional two phase handling works as follows
1. apply the requested changes to a clone of the model
2. detect if there are any conflicting changes to concurrent edits
3. if non-conflicting, merge the changes from 1. into the model, preserving any potential concurrent edits

The model only needs to be exclusively locked between 2. and 3. (contrary to the regular approach where teh model has to be exclusively locked during the full change cycle).

This feature is currently disabled by default. To enable it, enable the Maven profile "with-emfcompare" for "stardust-web-modeler" and "pepepr-test" (requiring an m2eclispe project refresh).

git-svn-id: http://emeafrazerg/svn/ipp/product/trunk/stardust/ui.web@59477 8100b5e0-4d52-466c-ae9c-bdeccbdeaf6b