[CRNT-33640]: Document Triage - Numerous FIxes
Clicking any collapsible button in the scanned document division as well as peroforming a mousedown on an image-stack will now
reset the selected page indexes. This addresses the issues where dragging top-level documents would result in an unwanted split as the selected
child documents werent cleared properly.

Created a  directive to force DOM elements to stretch to the bottom of the page and monitor numerous resize events of window and other DOM objects.
https://gist.github.com/mccainz/4bd9c6813e705baa8ce5
This addresses the issue with the scanned document panel not extending to the bottom of the page as well as the page viewer.

Added collapsible structure for the tree containers. Also added the ng-stretch-v directive to the startable container to maximize vertical space.
Fixed issue where staged process attachments were losing their document types on a tree refresh (startable only);
On drag from the scanned document division, the associated data is now passed to angular.copy so as to avoid referencing the original object.