[CRNT-33640]: Document Triage - Business Object Search Panel Directive
Created a directive which allows the user to search for business objects.
Initialization of business objects is baked into the directive via a service.
A single attribute is exposed on the directive to allow the user to register callbacks
to handle a user selecting a business object instance from the results table.