[CRNT-37662]: HTML5-Common Directives: sdTree directive indicates the wrong node in certain recursive use cases.
Added a srcScope property to the data object passed to the tree event callback function. This is goign to be a bit confusing in that srcScope will usually point to the scope of the tree node of interest, except in cases where the event is a tree-menu event. In that case srcScope actually indicates the scope of the item clicked in the tree menu. Rule of thumb is to leverage srcScope for all tree events other than menu events. In those cases one should refer to the data.nodeId property and retrieve the scope through the API (if scope is needed).

Bug: CRNT-37662