[CRNT-36274]: HTML5 - Drag and Drop is not working in Firefox & IE
Added href attribute to sdTree tree node template. This fixes an IE9 issue where drag would not work due to the absence of this attribute.

Added guard code to the sd-data-drag directive to assign the ngModel value to an empty string in the case where the ngModel value is undefined (which causes an error in the setData call of the DataTransfer API.

Bug: CRNT-36274