[CRNT-37840]: Support for badge icons in sdTree and sdAutoComplete directives.
Added nested icon structure to sd-tree-nodes.
Structure does not contain callbacks but does come pre loaded with the base badge and protrait classes. Tree node classes which need a badge structure will have to leverage CSS in order to specify the content of the badge bg and icons.
Parent icon is still dynamic of course.

Bug: CRNT-37840