[CRNT-37975]: Portals: Tree component to display a folder / file structure from the Document Repository.
Added levelOfDetail query parameter to each service call. This allows us to retrieve valid
information regarding the not yet acquired children of our target node. Added markup
to the directive template to properly calculate if a folder node is a leaf node based on the
new information returned from the server.

Bug: CRNT-37975