[CRNT-38329]:Benchmark Management: Benchmark can be saved with blank values
Very odd error. IE was reporting a script error from a comment line in the JS source of sdTree.
sdTree has comments such as "//@Param: myParam" in its source. For some reason this was causing a script error.
Most likely some odd encoding bug which was happening as the file was served from tomcat. The same file works fine in IE when used in a plnkr.
I have simply removed the "@" symbols in order to avoid the error.

Bug: CRNT-38329