[CRNT-38043]: Able to Edit the Rule set when "Publish" option is select to rule set context menu
One of the disable blocks was not using m_utils.jquerySelect and utilizing $() instead. This resulted in the selector not accessing the correct tab and thus the wrong UI elements for the disable attribute.

Parameter definition table was re-enabling UI on a click of the table. Modified code to add a click handler to the table to again disable UI on a click of the parameter definition table.

Bug: CRNT-38043