[CRNT-37905]: Web modeler: Model details get cut in model tree when scrollbar starts coming due to many models
Modified padding of the toolbar items so that they have enough room to display inline even with the vertical scrollbar active.
As the problem only manifested when the toolbar line wrapped to the next line this should be sufficient until such time as more toolbar items are added (if ever).

Bug: CRNT-37905