[CRNT-38434]: Authorization Manager View: Filter does not work as expected.
Fixed bad HTML caused by a merge.
Modified text filter so that an empty or undefined filter value will explicitly reset the tree filter via the treeAPI.
Added a message to indicate to the user when their filter returns zero matches.
Updated sdTree so that it exposes a resetFilter function.
Changed the treeAPI resetFilter function so that it does not collapse nodes.
Removed an orphaned bracket in the portal icon file caused by a bad merge.

Bug: CRNT-38434