[CRNT-38655]: Admin Portal - Participant Mangement : Participant Managment Tree should not allow drop operations on restricted nodes.
Modified sdTree.js so that it now contains a new recursive directive for dynamic tree generation. This was required to support dynamic droppable behaviors on recursive nodes. As a bonus the new recursive directive actually generates sd-tree-nodes.
Added logic to participant controller and view to support recursion factory and droppable expressions.
Added drop-over style to sd-tree css which matches the selected style.
Modified sdDrapDrop sd-data-drop so that it will add a drag-over class and remove it on drag-enter->drag-leave.
Modified the participant tree isLeaf logic to work correctly as before only users were being treated as leaves.

Bug: CRNT-38635