[CRNT-38693]: Participant Manager:Create User option not displayed by right click on any role under participants
The second level nodes in the tree did not have an sda-menu attribute defined on them, thus no second level would exhibit a menu. I have added the attribute which should fix the issue, assuming controller side logic for menu item generation is correct.
The logic for determining valid drop targets is based on a negative assertion and did not include entries for either DEPARTMENT or DEPARTMENT_DEFAULT.
Updated the logic to include these entries.

Bug: CRNT-38693