[CRNT-38925]: Browser freezes/crashes when opening the Participant Manager view in admin portal.
The participant tree is now lazy-compiled. Only when a tree node is expanded does the client incur the cost of compiling the angular directive which represents that node. Effectively this means that the performance of loading the initial tree is directly related to the number of nodes in the top level of the tree.  Also, when a node is closed that node is then removed. This should keep the tree very responsive regardless of the total count of nodes in the tree.

Bug: CRNT-38925