[CRNT-39394]: BCC -> Benchmark Definition View -> Remove vertical and horizontal scrollbars where possible
-Removed one horizontal scrollbar caused by a badly positioned header element as well as the small horizontal scrollbar associated with the model tree. Modified the css for the tree in the instance so that labels will be clipped with ellipsis at 145px in width. Also added sdaTitle attributes to each node so that on hover the node will display its entire name.
Regarding the scrollbar which appears underneath your bechmark definition once you add enough categories to force a scroll: this scrollbar will remain as it is too cumbersome to rely on the terminal bottom document scrollbar to reposition the panel. This was reviewed with Kyle Hargett who agreed that this scrollbar was acceptable.

Bug: CRNT-39394