[CRNT-40069]: Document Repository: Rename folder not working as expected, exception occurred
- Still seeing the same blur event migrating up multiple times causing the directives rename-commit event to fire twice.
Suppressing with the normal battery of prevent default etc doesn't seem to work. I have added code that inspects the timestamp of the event and in the case where it matches the previous blur timestamp I return from the callback logic with a NOOP. Assumption here is that user is not fast enough to rename two nodes on the scale of microseconds.

Bug: CRNT-40069