[CRNT-40556]: Control Center- Report Manager - View Version History table of a Report showing error
Initial Analysis:
The data initialization function for the version history directive does not currently set its referenced data (the data consumed by the internal dataTable directive) to an empty array in the case where no data is returned by its REST query. This results in the data table trying to initialize with an undefined object when it expects, at the least, to have an empty array.
Fix:
Added an else branch so that empty data can be recognized and the referenced structure set to an empty array, thus satisfying the minimum requirements for the data table.

Bug: CRNT-40556