[CRNT-40069]: Document Repository: Rename folder not working as expected, exception occurred
Event.timeStamp, which is being leveraged to filter out duplicate events, is bugged in FireFox in that it always reports zero. Amazingly, this bug has been open since 2004.

 https://bugzilla.mozilla.org/show_bug.cgi?id=238041

For this reason, all but the first rename-commit will appear as a duplicate to our code logic when using FireFox.
Modified code to use Date.now values to define a one second window within which we will ignore any successive rename-commit events within the sdTree.

Bug: CRNT-40069