[CRNT-40965] : Participant or Group creation: Issue with Date field
-The sdDateTime condensed directive uses an internal custom parser to determine validity. When placed within a form element the validity value of the directive can disable the forms submit button if the value of the directive is not valid. This is baked in behavior for form controls in Angular and not custom behavior implemented by the directive. In this case, sdDateTimeCondensed was reporting invalid when both of its values (date and time) are undefined. This corresponds to the initial state of the directive when not given a seed value by way of its ngModel. This prevented the Apply button from being clicked.

I have modified the internal ngModelController parser function so that having an undefined date part and an undefined time part will now be considered valid. It is important to note here that forms that rely on an actual value in a sdDateTimeCondensed directive will need to test external to the directives validator logic as it will no longer report undefined as invalid.

BUG: CRNT-40965